pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "setuptools_scm >= 8"]
build-backend = "setuptools.build_meta"

[project]
name = "crc-bonfire"
dynamic = ["version"]
description = "A CLI tool used to deploy ephemeral environments for testing cloud.redhat.com applications"
readme = "README.md"
license = { text= "MIT License" }
requires-python = ">=3.6"
authors = [
    { name = "Brandon Squizzato", email = "bsquizza@redhat.com" },
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Utilities",
]
dependencies = [
    "app-common-python>=0.1.6",
    "cached_property",
    "chardet<7.5,>=7.4",
    "click>=7.1.2",
    "gql[requests]>=3.5.0",
    "importlib-metadata; python_version<'3.8'",
    "importlib-resources; python_version<'3.9'",
    "junitparser",
    "multidict",
    "ocviapy>=1.6.0",
    "python-dotenv",
    "PyYAML",
    "requests>=2.33.0",
    "sh",
    "tabulate",
    "wait-for",
    "packaging",
    "truststore",
]

[project.scripts]
bonfire = "bonfire.bonfire:main_with_handler"
bonfire-mcp = "bonfire_mcp.server:main"

[tool.setuptools_scm]

[tool.setuptools.packages.find]
# package names should match these glob patterns (["*"] by default)
include = ["bonfire*", "bonfire_lib*", "bonfire_mcp*"]
# exclude packages matching these glob patterns (empty by default)
exclude = ["tests*"]
# to disable scanning PEP 420 namespaces (true by default)
namespaces = false

[project.optional-dependencies]
test = [
    "mock",
    "pytest",
    "pytest-mock",
    "requests-mock",
]
lib = [
    "kubernetes>=29.0.0",
    "jinja2>=3.1.0",
]
mcp = [
    "mcp>=1.0.0",
]

[project.urls]
Homepage = "https://www.github.com/RedHatInsights/bonfire"

[tool.ruff]
line-length = 100
indent-width = 4

